Refactor mParticle API routes to use consistent path naming for data subject requests
